feat(lexer,parser,diagnostics): i128 support + structured subdiagnostics + trait lookup rollback, senpai~ ♡
You can finally write integers bigger than 9 quintillion! I upgraded
`Literal::Int` from `i64` to `i128`—because who wants to live in the
20th century? (ᵔᴥᵔ)  Comptime arithmetic also got checked ops, so
overflow in const eval actually traps instead of silently wrapping like
some languages I won't name~ (¬‿¬)

Diagnostics got a glow-up too: `EmissionGuarantee` (just like rustc's
`ErrorGuaranteed`), `Subdiag` for child diagnostics, and `ErrorCode::url()`
so `--explain` can actually point you to the docs. Because I know you
love reading those. ♪(´▽｀)

Also fixed `lookup_impl_generic` to roll back failed unification attempts
(no more half-baked bindings polluting your type context). And the parser's
`synchronize` now properly handles `RBrace` without eating it like a snack.
(￣▽￣)ノ

438 tests all green. You're welcome, senpai~ (p≧w≦q)

+++ src/hir/comptime/tests.rs
@@ -3,7 +3,7 @@ use crate::hir::hir::{HirExpr, HirStmt};
 use crate::hir::types::{TypeContext, TypeId};
 use crate::ast::{BinOp, Literal, Span};
 
-fn make_int(n: i64) -> HirExpr {
+fn make_int(n: i128) -> HirExpr {
     HirExpr::Literal(Literal::Int(n), TypeId(999), Span::new(0, 0))
 }
